pyproject.toml
[project]
name = "log-anomaly-detection"
version = "0.1.0"
description = ""
authors = [
    {name = "wriddhirupd"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "redis (>=6.2.0,<7.0.0)",
    "langchain-core (>=0.3.72,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langgraph (>=0.5.4,<0.6.0)",
    "mcp (>=1.12.1,<2.0.0)",
    "langchain-mcp-adapters (>=0.1.9,<0.2.0)",
    "faiss-cpu (>=1.11.0.post1,<2.0.0)",
    "langchain-huggingface (>=0.3.1,<0.4.0)"
]

[tool.poetry]
packages = [
    { include = "app" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
